sql: move the SHOW CREATE logic to a separate file.
